RAG_demo/requirements.txt
fastapi
uvicorn
pydantic<3
langchain
langchain-community
langchain-text-splitters
langchain-google-genai
faiss-cpu
pypdf
pymupdf
python-dotenv
